requirements.txt
Django==2.2.16
pytz==2018.5
djangorestframework
markdown
django-filter
django-rest-passwordreset
